Adds notice that git is neccessary for degit
I ran the `npx degit ...` command on a new Win 10 machine with freshly installed node and had a degit/git error.
```
! could not find commit hash for master
```
It seems like degit uses git under the hood and therefore needs it to be installed (https://github.com/Rich-Harris/degit/issues/37).
I added a short pararagraph explaining this and also an alternative to using degit.
In my opinion the process of setting up a new project with new technologies should be as hassle-free as possible. This can easily throw newbies off before even really starting.

Secondly, I would also suggest to mention that `npm run dev` starts a **hot** development server which updates any saved `src` file changes immediatily visible in the browser. Would be nice to now that this is already set up by the svelte template, so that playing around with the code can get started ;)

At the very end of the article after the `App.svelte` is introduced, a recommended Visual Studio code extensions for syntax highlighting of svelte would be nice imo.